Improve error handling

+++ src/driver/path.sh
@@ -2,16 +2,26 @@ function bashelor.driver.path() {
 	local URL="$1"
 	local DIRECTORY="$2"
 
-	if [ -d "$DIRECTORY" ]
+	if [ ! -d "$URL" ]
 	then
-		bashelor.logger.log "=> Updating $(bashelor.logger.success $DIRECTORY)"
-		rm -rf "$DIRECTORY"
-	else
-		bashelor.logger.log "=> Installing $(bashelor.logger.success $DIRECTORY)"
+		bashelor.logger.error "   Could not install $DIRECTORY: Directory $URL not found"
+
+		return 2
 	fi
 
-	mkdir -p "$DIRECTORY"
+	if [ -d "$DIRECTORY" ]
+	then
+		bashelor.logger.log "=> Updating $(bashelor.logger.success ${DIRECTORY})"
+		rm -rf "$DIRECTORY"
+	else
+		bashelor.logger.log "=> Installing $(bashelor.logger.success ${DIRECTORY})"
+	fi
 
-	cp -r "$URL"/* "$DIRECTORY"
-	bashelor.logger.warning "   Installed $DIRECTORY"
+	mkdir -p $(dirname "$DIRECTORY")
+	cp -r "$URL" "$DIRECTORY"
+
+	if [ -d "$DIRECTORY" ]
+	then
+		bashelor.logger.warning "   Installed $DIRECTORY"
+	fi
 }
